Remove the dependency on the systemd package
The systemd package and journal logging does not offer a huge benefit over
over simply logging to standard output and letting systemd drop that into
the journal. Additionally, it makes greenwave more difficult to install by
requiring systemd headers to be present. Combine this with the desire to
deploy greenwave as a container in OpenShift, I'd say it's worth
dropping the module.

Signed-off-by: Jeremy Cline <jeremy@jcline.org>