Get "scenarios" working for greenwave.
I spent some time this afternoon with @mohanboddu and @adamwill
investigating use of greenwave for gating rawhide composes.

https://infrastructure.fedoraproject.org/cgit/ansible.git/commit/?id=f693b6de0d6566881adc189404176edb0bef6121

One thing we ran into was that the *requirements* for a rawhide compose
can't be expressed only in terms of testcase names.  They must be
specified in terms of both the testcase name *and* the scenario.

This patch gives greenwave the ability to do that.  A passing
testcaserule now optionally allows also specifying a required scenario
for that testcase.

I don't yet have a full list of testcase X scenario combinations, but
when I have one, I'll commit it to ansible.