Change publication logic.
- This is related to and likely conflicts with #103.
- This is an attempt to fix #104.

This changes our publication logic such that we only publish a message
*for every decision context and product-version combination* that we
know about instead of publishing *for every policy and product-version
combination*.

I think this is the right way to go since it is how other systems will
be calling greenwave.  They call with a *decision context* in their
question, not with a particular policy in mind.  We might have 10
different policies which all apply to the same context -- but it is the
context that the external system cares about.  Policies should be a more
or less internal implementation detail for greenwave.