fail the docs build on Sphinx warnings
We want Jenkins to tell us about this, if we accidentally introduce
problems in our docs that trigger a Sphinx warning.

But we *don't* want the RPM build to fail if there's warnings, because
we expect some from sphinxcontrib-issuetracker if it fails to talk on
the network when we build in Koji.