Announce decisions about specified sets of subject keys.
Fixes #92, for now.

I ran into this while talking through how we would do compose gating with
@mohanboddu.

The issue is (as described in other terms in #92), that greenwave doesn't know
what to announce if the subject isn't in the form of a simple item/type pair.
Compose decisions have just such a subject.

Here, we add a list of "interesting subjects" to the configuration.  As results
come in off the bus, we look to see if any of them might match up with an
interesting subject.  If they do, then construct that subject, and ask
greenwave if a decision would change, and publish a message (as we have done
before).

Maintaining a list of interesting subjects indefinitely will be a chore.
Elsewhere, @dcallagh has proposed giving greenwave some state so that it can
record all of the subject queries it has ever received, and use that to
calculate its announcements.