Factor in absence waivers.
Previously, we did work to retrieve waivers from waiverdb that were stored in
relation to the absence of a result.

However, the `check` method of our rule short-circuited itself if a result was
missing and it never even consulted the correctly retrieved waivers.

This adjusts that logic to account for waivers-of-absences.  It also adds a
test, confirming that it works.