Create a custom FedoraAtomicCi rule
The issue is that bodhi always provides an NVR but if the package is not
being considered by the Fedora Atomic CI pipeline, its corresponding git
repository isn't being clone and its `original_spec_nvr` value not
included in the `pipeline.package.ignore` fedmsg message.
In other words, relying on `pipeline.package.ignore` messages when
querying info about a particular NVR is never going to work.

With this rule we are specifying a list of packages of interest (to be
filled) and a test_case_name.
Every package that is part of the list of interest must satisfy this
test case, all the others are considered satisfying the rule.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>
Signed-off-by: Ralph Bean <rbean@redhat.com>