docs: replace Contribution Guidelines with Development Guide
I am much more interested in helping contributors get started developing
Greenwave, and less in prescribing what they must do if they want to
contribute.

This patch changes the Contribution Guidelines doc to a Development
Guide instead, and updates its advice to match the current state of
Greenwave. The sections are ordered from most important to least
important, for someone who wants to write a patch.

This also matches the approach we have in the WaiverDB docs, which
already has a Development Guide.