Support new and old resultsdb message format.
The new resultsdb format is in the latest release, but greenwave doesn't know
how to understand it.  The goal here is to seamlessly support both the old
format and the new format, so we can roll out resultsdb internally and have
greenwave just keep working.

Currently, everything still works in Fedora even though the new resultsdb is
already live, because resultsdb there still publishes its old message format in
addition to the new format.

I copied one of the original message tests to be a new "legacy" test verbatim,
to ensure that the old handling still works.