Introducing RemoteRule for extending policies
Packagers become responsible for defining extensions to the policy for
their specific packages defining a greenwave.yaml file in the dist-git
repos of packages.
Greenwave checks if the file exists, and, if it does it pulls it down,
loads it, and uses it to additionally evaluate the subject of the
decision.