Add retry decorator to backend-frontend requests.
This is an addendum to 251c008372a3.

There, a retry decorator was added to try and work around some DNS issues when
we can't even establish a connection.  We covered connections from
greenwave/web to resultsdb and waiverdb, but I forgot about greenwave's
connection from the backend to greenwave/web.  This covers that third case.

It gets tricky like the comments says because the backend doesn't have access
to the frontend configuration - no application context.  We should think about
some kind of unified config for a future change.