Rename greenwave.yaml to gating.yaml.
We're going to be asking packagers to put these files in their dist-git repos.
Rather than name the file after the system that expects it, name it after the
usecase with respect to the packagers' workflow.  The files defines criteria
for their package to pass various gating points, so call it `gating.yaml`.

:bike: :house_with_garden: