Backport "special relevance" from dist-git.
https://src.fedoraproject.org/rpms/greenwave/blob/3b12bdd78258ab1d63976b5beddc0b98bbb4b9b2/f/0002-Special-relevance.patch

See the commit message of that linked patch for the original logic going into this.

The problem now is that I left it as a dist-git patch for so long, it doesn't
apply nicely.  The API v2 stuff that @dcallagh is working on should let us
delete this work, but we need this upstreamed to simplify the release process
in the meantime.