run fedmsg consumers with a Flask app context
This lets us refer to current_app.config throughout the code, instead of
awkwardly having this thing where a piece of code can use
current_app.config if it was called through a Flask handler but
otherwise has to pass around an explicit config object.

The 'greenwave_cache' setting from the fedmsg config is dropped, because
we can just re-use the existing 'CACHE' setting from the Greenwave
config instead.