fix unit tests in RPM build
PR#199 broke the tests inside the RPM build, because now we are calling
create_app(), and it was unintentionally loading a production config
instead of a testing config.

The fix for that is really PR#192 but that was not taking effect because
the conftest.py file in the root directory was not being included in the
source tarballs.