Greenwave should consider that aliases exist
Greenwave was checking if the url.netloc of the build['source']
found in the koji/brew build is in the configured DIST_GIT_BASE_URL.
The problem is that the dist-git server can have an alias, and
this check is raising an error blocking the decision.
Removing the check for now... maybe in the future we can find a
smarter way to check that.