Fix wrong retrieving of gating.yaml file
For the RemoteRule feature the retrieving of the gating.yaml file
was wrong: Greenwave was using the subject_identifier (nvr) to
guess the pkg/container name to get the repo url for the gating.yaml
file. But this is not always right, we should use the source link in
the build received from koji/brew.