Remove Bodhi dependecy
Removes dependency on Bodhi - i.e. asking for all build from a Bodhi
update.

This removes cyclic dependency (Bodhi depends on Greenwave) and
simplifies the code.

Decision for bodhi_update no longer expands to include related
koji_build items from the Bodhi update. All builds have to be stated
explicitly in the "subject" field.

Decision change message for bodhi_update is no longer published if a
test results changes for a koji_build in the Bodhi update.

As an side effect, the formerly deprecated "subject" field (replaced
with "subject_identifier" and "subject_type") has to be used to query
for a decision on multiple koji_builds.

Fixes #298

Signed-off-by: Lukas Holecek <hluk@email.cz>