Remote policies not considered for decision change
Decision change message only respects policies configured locally on
the server and ignores RemoteRule rules.
If a result changes for a testcase for some Koji build and there is
an applicable policy on the server containing RemoteRule, Greenwave
should also emit decision change messages for policies defined in
gating.yaml in dist-git for the Koji build.