Support SCM URLs without the namespace.
It is possible to have Koji build from SCM URL without the namespace.
Currently, Greenwave disallows this which makes it unusable for
such builds.

In this commit, the `namespace` is set to an empty string in this
case and methods using namespace are updated to behave correctly
when empty string is passed to them as namespace.