Retrieve only latest results with verbose=True
When asking for a decision to the GW APIs with flag verbose=True,
Greenwave retrieves all matching results from ResulsDB. When those
results are *a lot*, the request hits timeout causing the all
decision call to fail. It would be just better to ask the lastest
results for that filtering, since we don't really care about the
old ones.