Omit comparing result_id values for decision change
When Greenwave receives new result message from ResultsDB it tries to
compare old decision (ignoring the new result) with new one (for all its
policies) so it can publish decision update message only when the
decision changed.

The new decision was seen as "changed" when any of its data differ from
the old decision. The problem is that decision data include result IDs
so it's always seen as "changed" if the new result is part of the new
decision.

Example:

    # Decision before new result is available:
    {
        "applicable_policies": [
            "taskotron_release_critical_tasks_for_stable"
        ],
        "policies_satisfied": true,
        "satisfied_requirements": [
            {
                "result_id": 27968725,
                "testcase": "dist.abicheck",
                "type": "test-result-passed"
            },
            {
                "result_id": 28015726,
                "testcase": "dist.rpmdeplint",
                "type": "test-result-passed"
            }
        ],
        "summary": "All required tests passed",
        "unsatisfied_requirements": []
    }

    # Decision after new result with ID=28015899 is available:
    {
        "applicable_policies": [
            "taskotron_release_critical_tasks_for_stable"
        ],
        "policies_satisfied": true,
        "satisfied_requirements": [
            {
                "result_id": 27968725,
                "testcase": "dist.abicheck",
                "type": "test-result-passed"
            },
            {
                "result_id": 28015899,
                "testcase": "dist.rpmdeplint",
                "type": "test-result-passed"
            }
        ],
        "summary": "All required tests passed",
        "unsatisfied_requirements": []
    }

With this change, the new decision is seen as "changed" when any of its
data changes *except* result IDs.

Fixes #395

Signed-off-by: Lukas Holecek <hluk@email.cz>