Check old decision before a specific time
The decision endpoint allows to pass results and waivers IDs lists to ignore
(ignore_result, ignore_waiver). These are used to compare the new decision
with older one. In case of multiple new results or waivers there could be a
race condition. This change introduces new parameters results_since and
waivers_since, used to determin the decision before these specific dates.
This solves the race conditions. ignore_result and ignore_waiver are not
used anymore to gather the old decision, but they are still parameters of the
API for backwards compatibility.