Add support for on-demand policies
The change consists in enhancing the /decision endpoint API to allow a new parameter "rules" that will allow the user to pass some rules. These rules will be immediately processes by Greenwave that will, “on demand”, check the decision (as usually querying ResultsDB and WaiverDB) for those rules and return a response.