Respond with a nicer error when the gating.yaml source is down
If you turn on the RemoteRule, greenwave will retrieve the `gating.yaml` from another source, which could be down or returning 500 for X or Y reasons.

In this situation, greenwave returns a 500 error without any explanations.

This change would have greenwave handle this situation a bit better and return with an error code some error message as to why it cannot proceed as expected.