Use better example decision request
The former example has incomplete "satisfied_requirements" and contains
improperly JSON-formatted response. It also contains waiver which is not
used to waived a missing/failed result and does not have a matching
product version.

New example is complete (but some unnecessary headers were removed),
properly formatted and contains a waiver that is used.

Signed-off-by: Lukas Holecek <hluk@email.cz>