Add topic to the message body
This seems silly, but it is here to appease a codepath for
legacy fedmsg-based consumers, notably: datanommer.

This codepath is to blame: https://github.com/fedora-infra/fedmsg/blob/develop/fedmsg/consumers/__init__.py#L254-L259

There, if there is no topic in the body, then the fedmsg base class
that prepares the message before it is passed onto the application's
`consume` method tries to be helpful by "massaging" what is assumed
to be a stomp message from a non-fedmsg based system into
a fedmsg-friendly message by wrapping a `msg: {` envelope around it.

This causes problems reported in EXDSP-1169.

Ideally, that fedmsg codepath should be updated to do this more
carefully. Adding a topic to the greenwave message body like in
this commit should be harmless to other consumers and should succeed
in working around the legacy fedmsg framework.