If we can't parse SCM URL from Koji, send an error result (#218) (#220)
If Greenwave can't parse the SCM URL for a Koji build (which can
happen if the build is run from a .src.rpm, which Koji admins
can do), it raises an exception in `_get_sub_policies`, which
ultimately results in the query returning 502. This doesn't give
a great experience in Bodhi - it'll just show no automated test
results and the gating status will be stuck at 'waiting' forever.
Using browser developer tools you can determine that it's getting
a 502 from Greenwave, but you can't tell why (we had to find that
out from greenwave's logs, in the real-world case here).

If we have Greenwave instead handle the exception and send a
response that includes an unsatisfied requirement indicating the
error, that should result in a better experience in Bodhi. The
gating status will be failed, test results will be shown, and the
UI will give at least some indication that there was an error
trying to retrieve remote policies. With developer tools you
should see the exact error in the Greenwave response JSON.

Signed-off-by: Adam Williamson <awilliam@redhat.com>