Remove subject_type attribute supports_remote_rule
Currently, setting `supports_remote_rule` or setting it to `false` means
that any `RemoteRule` (even with specified sources/URLs) in policies is
ignored for that subject type. In most cases, this is confusing
unexpected behavior.

The change is to drop the attribute completely and rely on presence of
`RemoteRule` in policies (it should not be present if the subject type
the policy uses does not support it).

JIRA: RHELWF-10561